Initial commit for #41
