+++ flask_server/blueprints/__init__.py
@@ -1,0 +1,3 @@
+from .main.views import blueprint
+from .auth.views import blueprint_auth
+from .api.views import blueprint_api
